feat: publish Kotlin service artifacts
This change is intended to generate the client and server definitions
for momento client protos. We should migrate MR2 and the control proxy
over to this, and we should consider migrating the Java SDK as well as
this should be making a java-shaped jar additionally. We might want to
publish these to jfrog as well for client consumption.