feat: Add _ to wire types for disambiguation with SDK Types (#29)
Some of our SDK response/request objects share the same name as wire types. Hence when customers try to import these classes the autocomplete shows multiple options and it may be confusing for customers to choose the correct name.

As a way to establish this pattern, I am proposing that all wire types should just be declared with `_` names. It will prevent future confusion as we add new messages as to why some messages have _ while other don't

From:
https://www.notion.so/momentohq/Request-Response-SDK-Names-6480de775fe84298aaa15eb91315d208