feat: Add proto declarations for dictionary. (#51)
* Add proto declarations for hashmap.

Add the declarations for HashGet, HashGetAll HashSet rpc's and the
associated request and response messages.

* Use list of key-value pairs to represent hash instead of mapping.

Previously I defined maps using bytes as keys and values. Protobuf does not support `bytes` for the keys or values in a
map type. To sidestep this I defined a new message type to represent a
key and value in a mapping from bytes->bytes. The mapping is therefore
a list of these key-value pairs.

* Prefer "Dict" noun over "Hash".

* Rename Dict (short name) to Dictionary.

* Bump rust client protobuf library version.