feat: Add exists service method and request/response messages.
`Exists` takes a list of keys as input and returns a list of
`bool`s. `results[i]` is `True` if `keys[i]` exists, else `False`.

We decided on returning a *list* as opposed to the count of items. By
doing this we can support rich iteration use cases on the client
side. For example:

```python
response = client.exists("my-cache", *keys)
if not response.all():
   log.info(f"Expected {len(keys)} to exist but got
   {response.num_exists()}")
   for missing_key in response.missing_keys():
       # Do things
```