chore: create separate gradle project for producing java proto artifacts
Currently, we have a single jar artifact in the "kotlin-messages"
gradle project which bundles both java and kotlin proto artifacts.
We are also publishing this to our internal codeartifact repo.

This commit is the first of a series to clean this up and get these
published to maven central so that users can build our OSS projects
and we can get rid of the submodules on the Java SDK.

This commit introduces a new java gradle project that builds a jar
that contains only the java proto artifacts.  Subsequent commits
will include:

* Publishing this java protos jar artifact to maven central.
* Removing the java protos from the kotlin jar.
* Removing the codeartifact stuff from the kotlin jar and publishing
  that jar to maven central.
* Updating the java SDK to get rid of the submodules and consume
  the jar from maven central.
* Updating any places in the service code where we are referring
  to the kotlin artifact, to use the maven central coordinates
  instead.