feat: apply name proposal to DictionaryGetAll (#70)
* feat: apply name proposal to DictionaryGetAll

In line with the name proposal, this renames DictionaryGetAll* to
DictionaryFetch.

* Use dictionary nomenclature and found/missing pattern.

We are using the term "field" to refer to the keys of the
dictionary. This should reduce confusion vs cache keys. Similarly we
use the term "items" to refer to the contents of a dictionary (cf set
elements).

This also applies the Found/Missing pattern to `_DictionaryGetResponse` and
`_DictionaryFetchResponse`.

* Rename `delete_fields` to be consistent with other APIs.