feat: add `DictionaryIncrement`
This adds the RPC, request, and response messages for
`DictionaryIncrement`. The response includes a `status` field to
indicate whether the operation was successful or not.

An increment operation (`DictionaryIncrement` or regular `Increment`)
may fail if the stored value does not parse to a number. Other errors
to consider including are integer overflow or integer underflow.