feat: add `DictionaryIncrement` (#88)
* feat: add `DictionaryIncrement`

This adds the RPC, request, and response messages for
`DictionaryIncrement`. The response includes a `status` field to
indicate whether the operation was successful or not.

An increment operation (`DictionaryIncrement` or regular `Increment`)
may fail if the stored value does not parse to a number. Other errors
to consider including are integer overflow or integer underflow.

* Rename `ECacheIncrementResult` values to avoid collisions

Protobuf enum values are siblings to the enum type, not children of
it. Because of this the `Ok` value collides with the `Ok` of
`ECacheResult`. Hence we rename the values.