feat: add "Ping" rpc, for use in ensuring connections are active (#100)
This commit adds a basic Ping RPC.  Empty request payload, empty
response payload.  This can be used to ensure that the client/server
connection is healthy, or, for clients that don't provide a mechanism
to eagerly establish the connection, this gives us a way to do it
without impacting cache metrics for the client.