feat: move ping RPC to separate service (#101)
The main cache service requires a "cache" header with a valid
cache name on every request.  This doesn't work well for the
Ping RPC because we won't necessarily know a cache name (and we
really don't need to know one) at the time we issue the request.

Rather than hacking a bunch of special case logic into the
already-fairly-complex auth interceptor for the cache service,
this moves the ping RPC to a separate service where we can
run without those requirements.