chore: Rename Add to SetIfNotExists
Although Add may be familiar to some users who have experience with other caches, SetIfNotExists will be easily understood by anyone because it is very as to what it does. Add requires people to read docs to be certain of what it does, SetIfNotExists does not. I'm going to change it to SetIfNotExists.