feat: Add protos for Integer type
Add protos for APIs for get, set, and increment on integer type.

This is called `Integer` rather than `Counter` because it can be assigned in addition to being incremented. It is `Integer` so that should we ever want to add support for floats, we can add `Number`.

Although this is not part of the protos, a few implementation details:
- If increment is called on an integer that does not yet exist, it will be initialized as 0 first.
- If increment is called without a provided value to increment by, it will default to incrementing by 1.
- The value to increment by can be negative (that is, `Increment` will do all the things the Redis `INCR`, `DECR`, `INCRBY`, and `DECRBY` commands do, in one API.)