feat: Add Increment API
Add proto definitions for rpc, request, and response for `Increment`.

`Increment` takes a key, an integer by which to increment by (can be negative to decrement), and a ttl. If the key does not already exist, it initializes it at 0.

The `Increment` operation maybe may fail if the stored value does not parse to a number. Currently this error will surface as an `Unavailable` error; in my opinion in order for this API to be usable we will want to make sure we surface the full error.