feat: Add Increment API (#106)
Add proto definitions for rpc, request, and response for `Increment`.

`Increment` takes a key, an integer by which to increment by (can be negative to decrement), and a ttl. If the key does not already exist, it initializes it at 0.

The `Increment` operation will fail if the stored value does not parse to a number. 

`Increment` does not take a `refresh_ttl` bool parameter because that only applies to datastructures - under the hood this is just a "unary" operation for `refresh_ttl` does not apply.