fix: generate rust files under `$OUT_DIR` instead of `src` (#108)
* fix: generate rust files under $OUT_DIR instead of src
* Use tonic::include_proto instead of using an explicit mod.rs file

Previously, the build script generated files under /src. This breaks
when the cargo registry is mounted read-only, and, in any case, is not
something that build scripts are supposed to do. The correct approach is
to generate files under $OUT_DIR, which is guaranteed to always be
writable and unique per crate/configuration.

This commit changes the build script to build generate files under
$OUT_DIR and then uses the standard include_proto! macro to put
them under a predictable module name.