feat: Specify whether cache methods are idempotent
Use the idempotency_level option to specify whether the cache methods
are safe, idempotent, or unknown. Safe methods have no side effects.
Safe and idempotent methods can be retried. By specifying idempotency
here we won't have to rely on per-sdk tables to know which methods are
safe to retry. Tested by running the Java sdk integration tests with
this as a dependency.