feat: custom message extensions (#152)
For dynamic inspection of messages, the inaugural extension is just
a retry semantic. The quality of retryability is  a little different
from idempotence. Each message may choose if we think it should be
automatically retried independently of whether or not it pedantically
matches a mathematical definition.

this also disables the strict tsc check due to the proto2 descriptors.proto,
which generates to typescript that has some implicit any strictness violations.