fix: apply relative import hack to all generated grpc code
The previous version patched the imports on the cache and control
client gRPC code using the hard-coded paths. Because we have since
added more gRPC services (auth), we would have had to add that
replacement to the list; however, we forgot to do this.

Because remembering to update this list is error prone, we change the
code to instead perform this replacement in *any* gRPC service
generated code. This was the original intended behavior, but less
error-prone going forward now.