feat: use poetry for packaging and generate type interfaces (#168)
In preparation for the python probotuf backwards compatiblity PR,
we found it difficult to install the version 3 and version 4 protobuf and
grpcio-tools dependencies with the previous setup.

To remedy this, we refactor the client-protos python to use poetry.
This simplifies building, publishing, and maintenance.

Also adds type hints to the generated code. Before the protoc
invocation neither generated nor packaged type hints for the client
protos. This made developing against them quite painful. With this we
both generate, package, and declare the typed library typed according
to PEP-484.

For testing, we:

built the package, pointed the Python SDK to the local package, and ran integration tests
in a scratch PR, dry ran the release process and validated proper version and artifact size vs local build