feat: support python-protobuf < 3.20
There is an incompatiblity between code generated with protoc linking
against `python-protobuf>=3.20` and the earlier versions. This means
that users pinned to `python-protobuf<3.20` cannot use our SDK.

This change compiles the protos using both versions, a pre-3.20
version of `grpc-tools` and version 4 (which is backwards compatible
through 3.20). We bundle both versions, but dynamically import the
correct one in the `momento_wire_types` package top-level init file.

Because of this, a small adjustment is needed to consuming code. We
have to slightly modify the import statements. See `run-protoc.sh` for
an example.