feat: support python-protobuf < 3.20 (#170)
There is an incompatiblity between code generated with grpcio-tools>=1.48.2
and earlier versions. The later versions of grpcio-tools use
python-protobuf>=3.20, which includes import statements that
earlier versions do not have. Users with code pinned to python-protobuf<3.20
cannot use our SDK. Worse, they will only find out with a runtime exception.

This change compiles the protos using both versions, a pre-3.20
version of grpc-tools and version 4 (which is backwards compatible
through 3.20). We bundle both versions, but dynamically import the
correct one in the momento_wire_types package top-level init file.

Because of this, a small adjustment is needed to consuming code. We
have to slightly modify the import statements. See run-protoc.sh for
an example.

To test, I have built and linked the python SDK to this version. I have made the appropriate changes there, and will open a PR once this is published.