feat: add detail to MVI list indexes RPC (#239)
Previously the list indexes RPC only returned a list of index
names. This PR modifies this to return a list of index detail
containing index name, num dimensions, and similarity metric.

We also factor out the similarity metric message from the
`CreateIndexRequest` since it is common to create index, list indexes,
and (in the future) describe index.

Because the service is in early beta we have carefully weighed
changing this and have agreed the earlier the better.