feat: migrate get RPC responses away from hit/miss
With filter expressions, the hit/miss pattern no longer makes sense
for the `get` RPC responses. Because a user could specify a general
filter expression not (necessarily) involving id's, we have no way of
specifying something as a "hit" or  "miss".

Because of this we will gradually migrate away from the hit/miss
pattern. We will support hits and fully specifying the result
temporarily. After the clients are migrated to using the new way we
will retire the hit/miss structure.