feat: deprecate old delete/get item request/response model
This PR deprecates two features of the previous request/response model
for `delete item batch`, `get item metadata batch`, and `get item
batch` RPCs:

(1) ids in the request: these RPCs now accept a general filter, which
can contain a set of ids. Because that behavior subsumes the old way,
we deprecate it.

(2) hit and miss pattern in get item response: because we no longer
always request items by id, we cannot use the hit and miss pattern
in the response. Instead we return a list of matching items.

Because the clients have already been updated to use the new
request/response model, this is not a breaking change.