chore: specify javascript grpc dependencies as peer dependencies (#315)
Because the JavaScript SDKs must also have the grpc
library (`@grpc/grpc-js` for node; `grpc-web` for web) installed, the
proto libraries here should declare those as peer dependencies, not
production dependencies.

In the previous form as production dependencies, both the SDK and the
protos libraries would have grpc as a dependency. In some situations
those dependencies do not deduplicate. Because we create grpc symbols
in the SDK to pass to objects from the protos library, those symbols
will not pass `instanceof` checks when deduplication fails.

By mandating grpc as a peer dependency, we avoid this
altogether. There will be one copy of grpc to load between the SDK and
the protos library: the one the SDK installs.